Refractoring. Deleted DEPRECIATED and FUTURE folders. Moved BackEnd to Game.MapTools
