+++ src/main/assets/Lore/lore.md
