+++ src/main/java/RoboRally/Game/MapTools/MapLoader.java
@@ -1,4 +1,4 @@
-package RoboRally.GUI.Screens;
+package RoboRally.Game.MapTools;
 
 import RoboRally.RoboRally;
 import com.badlogic.gdx.Gdx;
